api: move configuration file related struct to dedicate file
Move configuration file related struct to dedicate file.
