+++ api/src/config.rs
@@ -1,0 +1,557 @@
+// Copyright 2022 Alibaba Cloud. All rights reserved.
+// Copyright 2020 Ant Group. All rights reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+use std::collections::HashMap;
+use std::fs;
+use std::io::Result;
+
+use serde::Deserialize;
+use serde_json::Value;
+
+/// Configuration information for storage backend.
+#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
+pub struct BackendConfig {
+    /// Type of storage backend.
+    #[serde(rename = "type")]
+    pub backend_type: String,
+    /// Configuration for storage backend.
+    /// Possible value: `LocalFsConfig`, `RegistryConfig`, `OssConfig`.
+    #[serde(rename = "config")]
+    pub backend_config: Value,
+}
+
+/// Errors generated by/related to the API service, sent back through [`ApiResponse`].
+impl BackendConfig {
+    /// Create a new instance of `BackendConfig`.
+    pub fn from_str(backend_type: &str, json_str: &str) -> Result<BackendConfig> {
+        let backend_config = serde_json::from_str(json_str).map_err(|e| {
+            error!("failed to parse backend config in JSON string {:?}", e);
+            e
+        })?;
+
+        Ok(Self {
+            backend_type: backend_type.to_string(),
+            backend_config,
+        })
+    }
+
+    /// Load storage backend configuration from a configuration file.
+    pub fn from_file(backend_type: &str, file_path: &str) -> Result<BackendConfig> {
+        let file = std::fs::File::open(file_path).map_err(|e| {
+            error!("failed to open backend config file {}: {:?}", file_path, e);
+            e
+        })?;
+        let backend_config = serde_json::from_reader(file).map_err(|e| {
+            error!("failed to parse backend config file {}: {:?}", file_path, e);
+            e
+        })?;
+
+        Ok(Self {
+            backend_type: backend_type.to_string(),
+            backend_config,
+        })
+    }
+}
+
+/// Configuration information for localfs storage backend.
+#[derive(Clone, Default, Deserialize, Serialize)]
+#[serde(default)]
+pub struct LocalFsConfig {
+    /// Blob file to access.
+    #[serde(default)]
+    pub blob_file: String,
+    /// Dir to hold blob files. Used when 'blob_file' is not specified.
+    #[serde(default)]
+    pub dir: String,
+    /// Alternative dirs to search for blobs.
+    #[serde(default)]
+    pub alt_dirs: Vec<String>,
+}
+
+/// OSS configuration information to access blobs.
+#[derive(Clone, Default, Deserialize, Serialize)]
+#[serde(default)]
+pub struct OssConfig {
+    /// Enable HTTP proxy for the read request.
+    pub proxy: ProxyConfig,
+    /// Enable mirrors for the read request.
+    pub mirrors: Vec<MirrorConfig>,
+    /// Skip SSL certificate validation for HTTPS scheme.
+    pub skip_verify: bool,
+    /// Drop the read request once http request timeout, in seconds.
+    #[serde(default = "default_http_timeout")]
+    pub timeout: u32,
+    /// Drop the read request once http connection timeout, in seconds.
+    #[serde(default = "default_http_timeout")]
+    pub connect_timeout: u32,
+    /// Retry count when read request failed.
+    pub retry_limit: u8,
+    /// Oss endpoint
+    pub endpoint: String,
+    /// Oss access key
+    pub access_key_id: String,
+    /// Oss secret
+    pub access_key_secret: String,
+    /// Oss bucket name
+    pub bucket_name: String,
+    /// Oss http scheme, either 'http' or 'https'
+    #[serde(default = "default_http_scheme")]
+    pub scheme: String,
+    /// Prefix object_prefix to OSS object key, for example the simulation of subdirectory:
+    /// - object_key: sha256:xxx
+    /// - object_prefix: nydus/
+    /// - object_key with object_prefix: nydus/sha256:xxx
+    #[serde(default)]
+    pub object_prefix: String,
+}
+
+/// Container registry configuration information to access blobs.
+#[derive(Clone, Default, Deserialize, Serialize)]
+#[serde(default)]
+pub struct RegistryConfig {
+    /// Enable HTTP proxy for the read request.
+    pub proxy: ProxyConfig,
+    /// Skip SSL certificate validation for HTTPS scheme.
+    pub skip_verify: bool,
+    /// Drop the read request once http request timeout, in seconds.
+    #[serde(default = "default_http_timeout")]
+    pub timeout: u32,
+    /// Drop the read request once http connection timeout, in seconds.
+    #[serde(default = "default_http_timeout")]
+    pub connect_timeout: u32,
+    /// Retry count when read request failed.
+    pub retry_limit: u8,
+    /// Registry http scheme, either 'http' or 'https'
+    #[serde(default = "default_http_scheme")]
+    pub scheme: String,
+    /// Registry url host
+    pub host: String,
+    /// Enable mirrors for the read request.
+    pub mirrors: Vec<MirrorConfig>,
+    /// Registry image name, like 'library/ubuntu'
+    pub repo: String,
+    /// Base64_encoded(username:password), the field should be
+    /// sent to registry auth server to get a bearer token.
+    #[serde(default)]
+    pub auth: Option<String>,
+    /// The field is a bearer token to be sent to registry
+    /// to authorize registry requests.
+    #[serde(default)]
+    pub registry_token: Option<String>,
+    /// The http scheme to access blobs. It is used to workaround some P2P subsystem
+    /// that requires a different scheme than the registry.
+    #[serde(default)]
+    pub blob_url_scheme: String,
+    /// Redirect blob access to a different host regardless of the one specified in 'host'.
+    #[serde(default)]
+    pub blob_redirected_host: String,
+}
+
+/// Configuration information for blob cache manager.
+#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
+pub struct CacheConfig {
+    /// Type of blob cache: "blobcache", "fscache" or ""
+    #[serde(default, rename = "type")]
+    pub cache_type: String,
+    /// Whether the data from the cache is compressed, not used anymore.
+    #[serde(default, rename = "compressed")]
+    pub cache_compressed: bool,
+    /// Blob cache manager specific configuration: FileCacheConfig, FsCacheConfig.
+    #[serde(default, rename = "config")]
+    pub cache_config: Value,
+    /// Whether to validate data read from the cache.
+    #[serde(skip_serializing, skip_deserializing)]
+    pub cache_validate: bool,
+    /// Configuration for blob data prefetching.
+    #[serde(skip_serializing, skip_deserializing)]
+    pub prefetch_config: BlobPrefetchConfig,
+}
+
+/// Configuration information to create blob cache manager.
+#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
+pub struct FactoryConfig {
+    /// Id of the factory.
+    #[serde(default)]
+    pub id: String,
+    /// Configuration for storage backend.
+    pub backend: BackendConfig,
+    /// Configuration for blob cache manager.
+    #[serde(default)]
+    pub cache: CacheConfig,
+}
+
+/// Configuration information for a cached blob, corresponding to `FactoryConfig`.
+#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
+pub struct BlobCacheEntryConfig {
+    /// Identifier for the blob cache configuration: corresponding to `FactoryConfig::id`.
+    #[serde(default)]
+    pub id: String,
+    /// Type of storage backend, corresponding to `FactoryConfig::BackendConfig::backend_type`.
+    pub backend_type: String,
+    /// Configuration for storage backend, corresponding to `FactoryConfig::BackendConfig::backend_config`.
+    ///
+    /// Possible value: `LocalFsConfig`, `RegistryConfig`, `OssConfig`.
+    pub backend_config: Value,
+    /// Type of blob cache, corresponding to `FactoryConfig::CacheConfig::cache_type`.
+    ///
+    /// Possible value: "fscache", "filecache".
+    pub cache_type: String,
+    /// Configuration for blob cache, corresponding to `FactoryConfig::CacheConfig::cache_config`.
+    ///
+    /// Possible value: `FileCacheConfig`, `FsCacheConfig`.
+    pub cache_config: Value,
+    /// Configuration for data prefetch.
+    #[serde(default)]
+    pub prefetch_config: BlobPrefetchConfig,
+    /// Optional file path for metadata blobs.
+    #[serde(default)]
+    pub metadata_path: Option<String>,
+}
+
+/// Configuration information for blob data prefetching.
+#[derive(Clone, Debug, Default, Eq, Hash, PartialEq, Deserialize, Serialize)]
+pub struct BlobPrefetchConfig {
+    /// Whether to enable blob data prefetching.
+    pub enable: bool,
+    /// Number of data prefetching working threads.
+    pub threads_count: usize,
+    /// The maximum size of a merged IO request.
+    pub merging_size: usize,
+    /// Network bandwidth rate limit in unit of Bytes and Zero means no limit.
+    pub bandwidth_rate: u32,
+}
+
+/// Configuration information for file cache.
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct FileCacheConfig {
+    /// Working directory to store state and cached files.
+    #[serde(default = "default_work_dir")]
+    pub work_dir: String,
+    /// Deprecated: disable index mapping, keep it as false when possible.
+    #[serde(default)]
+    pub disable_indexed_map: bool,
+}
+
+impl FileCacheConfig {
+    /// Get the working directory.
+    pub fn get_work_dir(&self) -> Result<&str> {
+        let path = fs::metadata(&self.work_dir)
+            .or_else(|_| {
+                fs::create_dir_all(&self.work_dir)?;
+                fs::metadata(&self.work_dir)
+            })
+            .map_err(|e| {
+                last_error!(format!(
+                    "fail to stat filecache work_dir {}: {}",
+                    self.work_dir, e
+                ))
+            })?;
+
+        if path.is_dir() {
+            Ok(&self.work_dir)
+        } else {
+            Err(enoent!(format!(
+                "filecache work_dir {} is not a directory",
+                self.work_dir
+            )))
+        }
+    }
+}
+
+/// Configuration information for fscache.
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct FsCacheConfig {
+    /// Working directory to store state and cached files.
+    #[serde(default = "default_work_dir")]
+    pub work_dir: String,
+}
+
+impl FsCacheConfig {
+    /// Get the working directory.
+    pub fn get_work_dir(&self) -> Result<&str> {
+        let path = fs::metadata(&self.work_dir)
+            .or_else(|_| {
+                fs::create_dir_all(&self.work_dir)?;
+                fs::metadata(&self.work_dir)
+            })
+            .map_err(|e| {
+                last_error!(format!(
+                    "fail to stat fscache work_dir {}: {}",
+                    self.work_dir, e
+                ))
+            })?;
+
+        if path.is_dir() {
+            Ok(&self.work_dir)
+        } else {
+            Err(enoent!(format!(
+                "fscache work_dir {} is not a directory",
+                self.work_dir
+            )))
+        }
+    }
+}
+
+/// Configuration information for network proxy.
+#[derive(Debug, Clone, Deserialize, Serialize)]
+#[serde(default)]
+pub struct ProxyConfig {
+    /// Access remote storage backend via P2P proxy, e.g. Dragonfly dfdaemon server URL.
+    pub url: String,
+    /// Endpoint of P2P proxy health checking.
+    pub ping_url: String,
+    /// Fallback to remote storage backend if P2P proxy ping failed.
+    pub fallback: bool,
+    /// Interval of P2P proxy health checking, in seconds.
+    pub check_interval: u64,
+    /// Replace URL to http to request source registry with proxy, and allow fallback to https if the proxy is unhealthy.
+    #[serde(default)]
+    pub use_http: bool,
+}
+
+impl Default for ProxyConfig {
+    fn default() -> Self {
+        Self {
+            url: String::new(),
+            ping_url: String::new(),
+            fallback: true,
+            check_interval: 5,
+            use_http: false,
+        }
+    }
+}
+
+/// Configuration for mirror.
+#[derive(Clone, Deserialize, Serialize, Debug)]
+#[serde(default)]
+pub struct MirrorConfig {
+    /// Mirror server URL, for example http://127.0.0.1:65001.
+    pub host: String,
+    /// HTTP request headers to be passed to mirror server.
+    #[serde(default)]
+    pub headers: HashMap<String, String>,
+    /// Whether the authorization process is through mirror? default false.
+    /// true: authorization through mirror, e.g. Using normal registry as mirror.
+    /// false: authorization through original registry,
+    /// e.g. when using Dragonfly server as mirror, authorization through it will affect performance.
+    #[serde(default)]
+    pub auth_through: bool,
+    /// Interval of mirror health checking, in seconds.
+    #[serde(default = "default_check_interval")]
+    pub health_check_interval: u64,
+    /// Failure count for which mirror is considered unavailable.
+    #[serde(default = "default_failure_limit")]
+    pub failure_limit: u8,
+    /// Ping URL to check mirror server health.
+    #[serde(default)]
+    pub ping_url: String,
+}
+
+impl Default for MirrorConfig {
+    fn default() -> Self {
+        Self {
+            host: String::new(),
+            headers: HashMap::new(),
+            auth_through: false,
+            health_check_interval: 5,
+            failure_limit: 5,
+            ping_url: String::new(),
+        }
+    }
+}
+
+fn default_http_scheme() -> String {
+    "https".to_string()
+}
+
+fn default_http_timeout() -> u32 {
+    5
+}
+
+fn default_check_interval() -> u64 {
+    5
+}
+
+fn default_failure_limit() -> u8 {
+    5
+}
+
+fn default_work_dir() -> String {
+    ".".to_string()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{BlobCacheEntry, BLOB_CACHE_TYPE_BOOTSTRAP};
+
+    #[test]
+    fn test_blob_prefetch_config() {
+        let config = BlobPrefetchConfig::default();
+        assert!(!config.enable);
+        assert_eq!(config.threads_count, 0);
+        assert_eq!(config.merging_size, 0);
+        assert_eq!(config.bandwidth_rate, 0);
+
+        let content = r#"{
+            "enable": true,
+            "threads_count": 2,
+            "merging_size": 4,
+            "bandwidth_rate": 5
+        }"#;
+        let config: BlobPrefetchConfig = serde_json::from_str(content).unwrap();
+        assert!(config.enable);
+        assert_eq!(config.threads_count, 2);
+        assert_eq!(config.merging_size, 4);
+        assert_eq!(config.bandwidth_rate, 5);
+    }
+
+    #[test]
+    fn test_file_cache_config() {
+        let config: FileCacheConfig = serde_json::from_str("{}").unwrap();
+        assert_eq!(&config.work_dir, ".");
+        assert!(!config.disable_indexed_map);
+
+        let config: FileCacheConfig =
+            serde_json::from_str("{\"work_dir\":\"/tmp\",\"disable_indexed_map\":true}").unwrap();
+        assert_eq!(&config.work_dir, "/tmp");
+        assert!(config.get_work_dir().is_ok());
+        assert!(config.disable_indexed_map);
+
+        let config: FileCacheConfig =
+            serde_json::from_str("{\"work_dir\":\"/proc/mounts\",\"disable_indexed_map\":true}")
+                .unwrap();
+        assert!(config.get_work_dir().is_err());
+    }
+
+    #[test]
+    fn test_fs_cache_config() {
+        let config: FsCacheConfig = serde_json::from_str("{}").unwrap();
+        assert_eq!(&config.work_dir, ".");
+
+        let config: FileCacheConfig = serde_json::from_str("{\"work_dir\":\"/tmp\"}").unwrap();
+        assert_eq!(&config.work_dir, "/tmp");
+        assert!(config.get_work_dir().is_ok());
+
+        let config: FileCacheConfig =
+            serde_json::from_str("{\"work_dir\":\"/proc/mounts\"}").unwrap();
+        assert!(config.get_work_dir().is_err());
+    }
+
+    #[test]
+    fn test_blob_cache_entry() {
+        let content = r#"{
+            "type": "bootstrap",
+            "id": "blob1",
+            "config": {
+                "id": "cache1",
+                "backend_type": "localfs",
+                "backend_config": {},
+                "cache_type": "fscache",
+                "cache_config": {},
+                "prefetch_config": {
+                    "enable": true,
+                    "threads_count": 2,
+                    "merging_size": 4,
+                    "bandwidth_rate": 5
+                },
+                "metadata_path": "/tmp/metadata1"
+            },
+            "domain_id": "domain1"
+        }"#;
+        let config: BlobCacheEntry = serde_json::from_str(content).unwrap();
+        assert_eq!(&config.blob_type, BLOB_CACHE_TYPE_BOOTSTRAP);
+        assert_eq!(&config.blob_id, "blob1");
+        assert_eq!(&config.domain_id, "domain1");
+        assert_eq!(&config.blob_config.id, "cache1");
+        assert_eq!(&config.blob_config.backend_type, "localfs");
+        assert_eq!(&config.blob_config.cache_type, "fscache");
+        assert!(config.blob_config.cache_config.is_object());
+        assert!(config.blob_config.prefetch_config.enable);
+        assert_eq!(config.blob_config.prefetch_config.threads_count, 2);
+        assert_eq!(config.blob_config.prefetch_config.merging_size, 4);
+        assert_eq!(
+            config.blob_config.metadata_path.as_ref().unwrap().as_str(),
+            "/tmp/metadata1"
+        );
+
+        let content = r#"{
+            "type": "bootstrap",
+            "id": "blob1",
+            "config": {
+                "id": "cache1",
+                "backend_type": "localfs",
+                "backend_config": {},
+                "cache_type": "fscache",
+                "cache_config": {},
+                "metadata_path": "/tmp/metadata1"
+            },
+            "domain_id": "domain1"
+        }"#;
+        let config: BlobCacheEntry = serde_json::from_str(content).unwrap();
+        assert!(!config.blob_config.prefetch_config.enable);
+        assert_eq!(config.blob_config.prefetch_config.threads_count, 0);
+        assert_eq!(config.blob_config.prefetch_config.merging_size, 0);
+    }
+
+    #[test]
+    fn test_proxy_config() {
+        let content = r#"{
+            "url": "foo.com",
+            "ping_url": "ping.foo.com",
+            "fallback": true
+        }"#;
+        let config: ProxyConfig = serde_json::from_str(content).unwrap();
+        assert_eq!(config.url, "foo.com");
+        assert_eq!(config.ping_url, "ping.foo.com");
+        assert!(config.fallback);
+        assert_eq!(config.check_interval, 5);
+    }
+
+    #[test]
+    fn test_oss_config() {
+        let content = r#"{
+            "endpoint": "test",
+            "access_key_id": "test",
+            "access_key_secret": "test",
+            "bucket_name": "antsys-nydus",
+            "object_prefix":"nydus_v2/"
+        }"#;
+        let config: OssConfig = serde_json::from_str(content).unwrap();
+        assert_eq!(config.scheme, "https");
+        assert!(!config.skip_verify);
+        assert_eq!(config.timeout, 5);
+        assert_eq!(config.connect_timeout, 5);
+    }
+
+    #[test]
+    fn test_registry_config() {
+        let content = r#"{
+	    "scheme": "http",
+            "skip_verify": true,
+	    "host": "my-registry:5000",
+	    "repo": "test/repo",
+	    "auth": "base64_encoded_auth",
+	    "registry_token": "bearer_token",
+	    "blob_redirected_host": "blob_redirected_host"
+        }"#;
+        let config: RegistryConfig = serde_json::from_str(content).unwrap();
+        assert_eq!(config.scheme, "http");
+        assert!(config.skip_verify);
+    }
+
+    #[test]
+    fn test_localfs_config() {
+        let content = r#"{
+            "blob_file": "blob_file",
+            "dir": "blob_dir",
+            "alt_dirs": ["dir1", "dir2"]
+        }"#;
+        let config: LocalFsConfig = serde_json::from_str(content).unwrap();
+        assert_eq!(config.blob_file, "blob_file");
+        assert_eq!(config.dir, "blob_dir");
+        assert_eq!(config.alt_dirs, vec!["dir1", "dir2"]);
+    }
+}
